refactor: consolidate chatretry into single file pair
