--- coderd/chatd/chatretry/backoff_test.go
@@ -1,38 +1,0 @@
-package chatretry_test
-
-import (
-	"fmt"
-	"testing"
-	"time"
-
-	"github.com/coder/coder/v2/coderd/chatd/chatretry"
-)
-
-func TestDelay(t *testing.T) {
-	t.Parallel()
-
-	tests := []struct {
-		attempt int
-		want    time.Duration
-	}{
-		{0, 1 * time.Second},
-		{1, 2 * time.Second},
-		{2, 4 * time.Second},
-		{3, 8 * time.Second},
-		{4, 16 * time.Second},
-		{5, 32 * time.Second},
-		{6, 60 * time.Second},  // Capped at MaxDelay.
-		{10, 60 * time.Second}, // Still capped.
-		{100, 60 * time.Second},
-	}
-
-	for _, tt := range tests {
-		t.Run(fmt.Sprintf("Attempt%d", tt.attempt), func(t *testing.T) {
-			t.Parallel()
-			got := chatretry.Delay(tt.attempt)
-			if got != tt.want {
-				t.Errorf("Delay(%d) = %v, want %v", tt.attempt, got, tt.want)
-			}
-		})
-	}
-}
